+++ source/Classes/011.SqlResourceBase.ps1
@@ -67,6 +67,7 @@ class SqlResourceBase : ResourceBase
             $connectSqlDscDatabaseEngineParameters = @{
                 ServerName   = $this.ServerName
                 InstanceName = $this.InstanceName
+                ErrorAction  = 'Stop'
             }
 
             if ($this.Credential)